@nanosoldier 2023-05-13
[Your benchmark job](https://github.com/JuliaLang/julia/pull/49664#issuecomment-1546163030) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/NanosoldierReports/blob/master/benchmark/by_hash/fe92e89_vs_d55314c/report.md).